requirements.txt
discord
pillow
moviepy
async